Move the query context out to its own file and rename to "access" context as it now also has write support.
Move the writing functions out of query.c and in to write.c
