+++ src/dnmd/CMakeLists.txt
@@ -1,12 +1,14 @@
 set(SOURCES
+  access.c
   bytes.c
+  deltas.c
+  editor.c
   entry.c
   query.c
   streams.c
   tables.c
   sig.c
-  editor.c
-  deltas.c
+  write.c
 )
 
 set(HEADERS
